refactor: move icons folder to components
