+++ src/components/icons/BagOutline.tsx
