+++ src/components/icons/Photo.tsx
