+++ src/components/icons/StackFill.tsx
