+++ src/components/icons/StackOutline.tsx
