+++ src/components/icons/Stars.tsx
