+++ src/components/layout/navbar/mobile/index.tsx
@@ -1,10 +1,10 @@
 'use client'
+import BagFillIcon from '@/components/icons/BagFill'
+import HomeIcon from '@/components/icons/Home'
+import MessageFillIcon from '@/components/icons/MessageFill'
+import MoreOutlineIcon from '@/components/icons/MoreOutline'
+import UserFillIcon from '@/components/icons/UserFill'
 import { useOnClickOutside } from '@/hooks/useOnClickOutsite'
-import BagFillIcon from '@/icons/BagFill'
-import HomeIcon from '@/icons/Home'
-import MessageFillIcon from '@/icons/MessageFill'
-import MoreOutlineIcon from '@/icons/MoreOutline'
-import UserFillIcon from '@/icons/UserFill'
 import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 import { ComponentProps, useRef, useState } from 'react'
